refactored roll method and added directions and default speed
