+++ SpecRunner.html
@@ -12,12 +12,12 @@
   <script src="lib/jasmine-2.4.0/boot.js"></script>
 
   <!-- include source files here... -->
-  <script src="src/sphero.js"></script>
+  <script src="src/movementCommands.js"></script>
   <!-- <script src="src/leapMotion.js"></script> -->
 
   <!-- include spec files here... -->
   <!-- <script src="spec/leapMotionSpec.js"></script> -->
-  <script src="spec/spheroSpec.js"></script>
+  <script src="spec/movementCommands.js"></script>
 
 </head>
