+++ src/leapMotion.js
@@ -1,0 +1,3 @@
+function app (MovementCommands){
+  
+}
